requirements.memory.txt
# Requirements for Memory Intelligence Layer

# Core dependencies
langfuse>=2.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation>=0.41b0

# Neo4j for Graphiti integration
neo4j>=5.0.0
py2neo>=2021.2.3

# API framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# OpenAI for LLM operations
openai>=1.0.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
httpx>=0.25.0
tenacity>=8.2.0

# Async support
aiofiles>=23.0.0
asyncio>=3.4.3

# Monitoring
prometheus-client>=0.19.0

# Testing (optional, for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
